backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.2
pandas==2.1.3
num2words==0.5.13
joblib==1.3.2
Pillow==10.1.0
python-decouple==3.8
SpeechRecognition==3.10.0
pydub==0.25.1
